print.obsSens(obsSens)	R Documentation

Print and Summarize ObsSens objects.

Description

Print a nice table of the results from a sensitivity analysis on observational data.

Usage


print.obsSens(x, ...)

summary.obsSens(object, digits=3, ...)

print.summary.obsSens(x, ...)

Arguments

x 	 An obsSens object or summary.obsSens object.

object 	 An obsSens object.

digits 	 Passed to format.

... 	 Additional arguments passed on to print methods for matrix/array.

Details

These functions print and summarize the results of the obsSens functions. The only difference in the print and summary methods is the return value (the information printed to the screen is the same). The print function returns a copy of the original object. The summary function returns a matrix/array of character strings corresponding to what is printed that can hopefully be used along with other functions (latex, xtable, odfTable, ...) for typesetting results in other formats.

Value

The print method returns a copy of the original object. The summary method returns a character array/matrix with the coefficients and confidence intervals to be passed to other typesetting/printing functions.

Author(s)

Greg Snow greg.snow@imail.org

Examples


# Recreate tables from above references

obsSensCCC( log(23.1), log(c(6.9, 77.7)), g0=c(2,6,10),
  p0=seq(0,.5,.1), p1=seq(0,1,.2) )

obsSensSCC( log(1.21), log(c(1.09,1.25)),
  p0=seq(0,.5,.1), p1=seq(0,1,.1), g0=3 )

obsSensCNN( log(1.14), log(c(1.10,1.18)),
  rho=c(0,.5, .75, .85, .9, .95, .98, .99),
  gamma=seq(0,1,.2), sdx=4.5 )

[Package obsSens version 1.0 Index]